// start-microfrontends/apps/se/src/app/application/tramites/110221/models/peru-certificado.model.ts
export interface AccionBoton {
  /**
   * @property {string} accion
   * @description
   * Acción que se debe realizar (por ejemplo, avanzar o retroceder en el wizard).
   */
  accion: string;

  /**
   * @property {number} valor
   * @description
   * Valor asociado a la acción, generalmente el índice del paso al que se debe navegar.
   */
  valor: number;
}